+++ internal/dto/person_dto.go
@@ -12,10 +12,6 @@ type PersonCreate struct {
 	Documents  []Document `json:"documents"`
 }
 
-type PersonID struct {
-	ID int `json:"id"`
-}
-
 type PersonUpdate struct {
 	ID         int    `json:"id"`
 	FirstName  string `json:"firstName"`
@@ -28,4 +24,8 @@ type PersonGet struct {
 	FirstName  string `json:"firstName"`
 	SecondName string `json:"secondName"`
 	MiddleName string `json:"middleName"`
+}
+
+type PersonDelete struct {
+	ID int `json:"id"`
 }
