+++ justification.md
@@ -14,36 +14,6 @@
 ### 1.1 Ручки
 Сейчас реализован CRUD по требованиям. На "/person" привязано четыре метода.
 
-[GET] - для получения пользователя по ID
-
-Если в json-структуре не будет поля ID, то вернется 400ый статус
-
-Пример запроса который нужно отправить:
-
-`{"id":1}`
-
-[POST] - для создания пользователя
-
-Пример запроса, котрый нужно отправить:
-
-`{ "firstName":"Ivan", "secondName":"Ivanov","middleName":"Ivanovich", "documents":[{"id":1, "name":"Справка"},{"id":2, "name":"Документ"}]}`
-
-[PUT] - для обновления пользователя
-
-Если не указан id, то вернется 400ый статус
-
-Пример запроса, котрый нужно отправить:
-
-`{"id":1, "firstName":"Ivan", "secondName":"Ivanov","middleName":"Ivanovich"}`
-
-[DELETE] - для удаления пользователя
-
-Если не указан id, то вернется 400ый статус
-
-Пример запроса, котрый нужно отправить:
-
-`{"id":1}`
-
 ### 1.2 Внедрение зависимостей
 На данный момент в "app.go" я все прописываю руками. В идеале, при масштабируемости (да и в целом желательно), использовать wire. На данный момент wire не добавлен, но система разработана так, что добавить его не проблема
 
@@ -93,6 +63,3 @@ query := db.Query("actors").Sort("id", false)
 
 ### 2.6 Потоковая обработка документов
 Это является самой интересной для меня задачей. Сейчас думаю, что буду реализовывать через каналы, но изучаю разные методы для реализации этой задачи
-
-### 2.7 Докер-образ
-Под конец всех исправлений хочется, чтобы проект был в Dockerfile, чтобы его можно было легко развернуть с тем зависимостями, на которых он разрабатывался
