+++ readme.md
@@ -26,32 +26,16 @@
 
 ## Использование API
 
-- [GET] - для получения пользователя по ID
+- [GET] api/v1/person/{id} - для получения пользователя по ID
 
-    Если в json-структуре не будет поля ID, то вернется 400ый статус
-
-    Пример запроса который нужно отправить:
-
-    `{"id":1}`
-
-- [POST] - для создания пользователя
-
-    Пример запроса, котрый нужно отправить:
+- [POST] api/v1/person - для создания пользователя
 
     `{ "firstName":"Ivan", "secondName":"Ivanov","middleName":"Ivanovich", "documents":[{"id":1, "name":"Справка"},{"id":2, "name":"Документ"}]}`
 
-- [PUT] - для обновления пользователя
-
-    Если не указан id, то вернется 400ый статус
+- [PUT] api/v1/person/{id} - для обновления пользователя
 
     Пример запроса, котрый нужно отправить:
 
-    `{"id":1, "firstName":"Ivan", "secondName":"Ivanov","middleName":"Ivanovich"}`
+    `{"firstName":"Ivan", "secondName":"Ivanov","middleName":"Ivanovich", "documents":[{"id":1, "name":"Справка"},{"id":2, "name":"Документ"}]}`
 
-- [DELETE] - для удаления пользователя
-
-    Если не указан id, то вернется 400ый статус
-
-    Пример запроса, котрый нужно отправить:
-
-    `{"id":1}`
+- [DELETE] api/v1/person/{id} - для удаления пользователя
